Test repository deleting transaction including reference changes
We're currently not testing whether repository deletions succeed
if the transaction contains some reference changes. Add a test for
this and modify the logic enough to have the test pass.

We don't actually have any transactions that perform other changes
with repository deletions in the same transactions. This is one case
which our checks don't handle but we also don't handle the other
combinations. This is a more general problem with our conflict checks
as they currently expect only certain shaped transactions. We aim
to generalize the conflict checking later so arbitrary transactions
can be conflict checked correctly. For now, add this test to cover
one of these cases.