Allow reftable job to fail
reftable job started failing due to an upstream change in Git that
changed the output of some commands. The change was recently merged
into Git's master branch which the 'reftable' job is using. The
offending commit is ac16f55697211ae4ef557706a098b8ed34a7fda6. The
change is triggering pipeline failures in Gitaly and blocking
merges of MRs. For now, allow failures of the reftable job until
the issue is fixed.