Add helper for skipping tests if Git version is less than expected
We're soon introducing new functionality that requires more recent
version of Git than our minimum. As the functionality is patched
in the bundled Git distribution and hasn't been yet released, we
can't simply bump the minimum version. The tests relying on the
new functionality will fail if the Git version doesn't contain the
expected patched. Add a helper to skip tests if the Git version is
less than expected. We'll use this to skip tests running into this
problem later.