Apply key-value operations from log entry
We soon need to store key-value data in the partitions along with
repository data. This data will for example contain a mapping of which
repositories the partition contains. The key-value data needs to also
go through the write-ahead log to ensure it can be replicated and
recovered on failures. This commit extends the TransactionManager to
apply the logged key-value operations as well when applying the log
entry.