praefect: Stop logging when repo does not have alternate
With 5aa03dd6, Praefect starting reconciling object pool state during
scheduled replication jobs. This requires checking the primary replica
to see if it is linked to an object pool and ensuring the state matches
the secondary replica.

The `GetObjectPool` RPC is used to get the object pool from a
repository. If there is no object pool, an empty response is returned
with no error. Despite this being perfectly normal, a message is logged
indicating that no alternates file exists.

Prevent this message from being logged because it seems to indicate that
there is a problem when there is none.