Make TestPartitionAssigner a table-driven test
TestPartitionAssigner is somewhat difficult to read at the moment
as there are multiple test cases in the same body. It's difficult
to see which state is relevant for which assertion. Change it to
a table-driven test which exercises specific scenarios. This way
it's easier to extend the tests later as well.