Fix pool path in TestOptimizeRepository
TestOptimizeRepository is using an incorrect object pool path in
tests with an object pool. It's writing the absolute path of the
alternate's object directory in the 'objects/info/alternates' file
instead of the relative path from the object directory of the linked
repository. Fix the issue and write the relative path from the linked
repository's object directory to the alternate repository's object
directory.