Add missing RPC to TestCompleteForkCreationFlow
When Rails creates an fork, it first issues a RepositoryExists
request to it. Our TestCompleteForkCreationFlow is not currently
sending a similar request which led to missing in our test suite
that the current flow does not work with Rails. Now that we've
fixed the issue, add the missing RPC invocation to the test.