Send a partitioning hint with CreateFork from Praefect
Praefect does not rewrite the source repository in CreateForkRequest.
This is because Gitaly expects to have the original non-rewritten
repository as it will fetch the source repository by going through
Praefect's API again. For Praefect to route the request correctly the
repository needs to be in its original form.

This creates a problem with transactions. Gitaly needs to place the
source repository and the new fork in the same partition in order
to link them both to the same pool later. With Praefect in use, the
repositories are stored on Gitaly at the rewritten paths generated
by Praefect. Gitaly doesn't thus know from the CreateForkRequest
which repository on the disk is the source repository that the new
fork should be partitioned with. To solve this, this commit sends a
partitioning hint from Praefect when routing CreateFork calls. The
hint is the source repository's path as it would be rewritten by
Praefect and stored on Gitaly. Gitaly uses this hint to locate the
source repository and ensure the fork ends up in the same partition
with it.