Add helpers for passing partitioning hints via gRPC metadata
Gitaly generally relies on the additional repository in the request
to determine if the request's target repository needs to be partitioned
with another repository. In some cases we don't have an additional
repository in the request but we know at the time of sending the request
that the target should be partitioned with another repository. An example
of this is when an object pool is first replicated to a Gitaly node, and
afterwards its members. As we're replicating the member, we know it should
join the pool and that we'll link it later. The ReplicateRepository request
itself doesn't contain enough information to determine this though. The
solution is to send Gitaly a partitioning hint when creating the repository
that needs to be partitioned with another one. That way we can eventually
then link the repositories together.

This commit adds helpers for passing partitioning hints via gRPC metadata.