lint: Don't use viper global state for config
We have been relying on `viper` pulling in the parsed config into its
global `config` before our custom linter is loaded. With `golangci-lint`
v1.57.0 this behavior changed and the config is now empty at the time
of loading.

Remove our custom `viper` logic to capture the arguments and instead
use `mapstructure` to unmarshal the provided `any` config to a useable
struct.