Make SetFullPath no-op in WAL transaction
SetFullPath RPC writes the repository's full path into the git config.
This is not supported by WALtransaction. That RPC was scheduled for
removal in some next versions. Until then, we should mark that RPC as a
no-op when being executed in the context of a WAL transaction.

This commit returns Unimplemented status code if the RPC handler detects
an active WAL partition manager. From the definition, Unimplemented
status code is for "The operation is not implemented or is not
supported/enabled in this service". That's a perfect status code for
this purpose.