Remove SetFullPath and FullPath RPCs
Gitaly's SetFullPath is used by Rails to store the repository's project
path inside the repository's git config. This is mainly done to have
some human-readable identifier in the repository should the database
mappings from the hashed path to the project path be lost.

We've concluded that we don't want to keep supporting this
feature:
* It's better to push users to maintain proper backups. With proper
  backups, one would always have the project path backed up along
  with the rest of the database.
* The full path in the repository can be stale and out of sync with
  the project path in Rails DB. It shouldn't be relied on.

So, we planned to remove SetFullPath RPC, FullPath RPC, as well as
the corresponding `gitlab.fullpath` config. That work was tracked in
this epic: https://gitlab.com/groups/gitlab-org/-/epics/8953. All the
callers were removed in this MR:
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/147651. This
commit removes the implementations of SetFullPath and FullPath RPCs.