Remove `gitlab.fullpath` from git config in repositories
This Git config was set to denote the project's full path of a
repository on Gitaly side. We scheduled the removal this feature in
this epic:
https://gitlab.com/groups/gitlab-org/-/epics/8953.

After the corresponding RPCs are removed, we need to clean up the
unused `gitlab.fullpath` git config in each repository. To do so, this
commit implements a stale data cleaning task. This task is triggered
as a part of the periodic housekeeping scheduler or by the
OptimizeRepository RPC.