conflicts: Treat `ErrUnexpectedDelimiter` as client side errors
Whenever a `ErrUnexpectedDelimiter` is encountered, we raise an internal
error. However, this error is only raised when the conflicts aren't
fully resolved. This generally means that the client missed resolving
all clients. This is not an internal error, rather this is a client side
error, since the client failed to resolve all conflicts. So let's catch
and treat such errors as `InvalidArgument` errors under GRPC.