diff: Support rename detection in `FindChangedPaths` RPC
The `FindChangedPaths` RPC is used to detect file changes via
`git-diff-tree(1)`. The `--no-renames` option is always used with this
RPC and thus detection of file renames is disabled. As part of the new
diff architecture, file renames must be detected to enable paginated
diffs via the `DiffBlobs` RPC and compute diffs between renamed files as
expected.

To facilitate this, the `find_renames` option is added to the
`FindChangedPaths` RPC request to toggle rename detection.