Remove ApplyGitattributes from WAL exclusion list
That RPC writes out the attributes file which is not supported with
WAL. The RPC is in the process of being removed. At this point, it
is essentially a no-op [1]. The RPC removes `info/attributes` file
of a repository when triggered. We don't use that file anymore. We
will have a housekeeping task that removes those files. This task
will run for all repositories before WAL is enabled.

So, we can simply remove it from the WAL exclusion list. Afterward,
the removal is executed in a transaction but the result is never
collected, meaning no changes are applied to the target repository.

[1]: https://gitlab.com/gitlab-org/gitaly/-/merge_requests/6573