Makefile: bump Git version to v2.44.0.gl2
We have recently tried to roll out Git v2.44.0 to production systems but
discovered that it breaks object reuse in repositories with a multi-pack
index. This issue led to user-visible errors as well as object reuse
being disabled completely when serving fetches.

We have fixed the issue upstream and backported the fix to our custom
Git v2.44.0.gl2. Upgrade to that version to unblock the rollout.