conflicts: Use `NewPath` for most of conflict resolution
In the previous commit, we added a test to capture the issue of
conflicts around files which are moved/renamed, we also dived into the
cause of the issue.

Let's fix the issue by simply using the `NewPath` for conflict
resolution. This also fixes the test which was added in the previous
commit.