Support skipping TransactionManager tests with macOS
Linux supports file system paths with invalid UTF-8 where as macOS
does not. Git allows non-UTF-8 reference names which may be written
out as file names if the reference is stored as a loose reference.
We'll soon add support for this, so we need a way to exclude the tests
from running on macOS.