Support non-UTF-8 file system paths in logging protocol
Linux allows for non-UTF-8 file system paths. Our logging protocol
is currently storing paths as strings which in protocol buffers are
expected to be UTF-8. To allow for non-UTF-8 paths, convert the
protocol buffer types to bytes and adjust the code around it.

Tests covering the UTF-8 support exclude macOS as macOS enforces
UTF-8 in paths.