Don't enable Git trace2 logging with all other feature flags
Gitaly has an environment variable to enable all feature flags. This
is used in tests by Rails to enable all of Gitaly's feature flags.
This causes the Git trace2 logging feature flag to be also enabled
leading to spamming the test logs with Git traces. Don't enable the
trace logging with the environment variable.