Don't perform no-op reference updates
TransactionManager is currently incorrectly recording updates of
no-op reference updates that target a packed reference. When we are
recording the reference changes, we expect that a reference is
written out as a loose reference if it is updated. If the update is
a no-op and the reference exists in a pack, this leads to an error
as there is no new reference file to stage.

Fix this by dropping no-op reference updates. We'll verify the update
by running it through update-ref but don't attempt to record its writes.
There are no writes to record after all since the update is a no-op.