backup: Assert sink by type rather than error
Currently we validate that we have the correct storage sink type by
performing an invalid reflection and confirming that the error contains
an expected string tangentially related to the backing storage. This is
a bit unintuitive and these errors may not be considered stable by the
gocloud maintainers.

Convert this to confirm that we can call the `As` function to convert
the sink to the bucket type associated with a given backend.