backup: replace FilesystemSink with fileblob
Our FilesystemSink has some subtle differences in behavior compared to
the gocloud implementations. For instance, creating a new file and then
canceling the context will leave an empty file, vs no change for gcloud.
Gocloud offers an equivalent `fileblob` backend that we can leverage and
save the effort of maintaining our own.

Replace FilesystemSink with the fileblob sink for local filesystem
operations. This does require us to adjust bundle tests to expect shared
directory and file permissions, rather than private, as fileblob does
not make this configurable.

We set the `NoTempDir` option as otherwise we would require `/tmp` and the
target directory be on the same filesystem. Many customers will be
saving backups to a mounted volume, and `/tmp` is commonly configured as
a RAMdisk.

Closes #5879