Don't error out when deleting a non-existent pool with transactions
When transactions are enabled, Gitaly returns a not found error when
attempting to delete a non-existent object pool. Without transactions,
Gitaly returns a successful response. This was missed as the test case
testing non-existent pool deletion is actually deleting a pool that
exists.

Fix the broken test and adjust behavior with transactions to return
the same successful response as without. Arguably returning a not
found would be a better interface this will fix Rails pipeline failures
with transactions for now.