Fix Gitaly hanging after cancelled InfoRefs request
When we have an info ref cache miss and the request context is
cancelled when `missFn` [1] is being executed, we could end up in a
situation where the command is blocked trying to write/copy the last bytes of
stdout to a pipe[2] that's not being read from (because this goroutine[3]
has exited at this point after the context was cancelled).

Discarding the remaining bytes [4] doesn't work because all reads from
the `TeeReader` are matched with writes to the writer it writes to, in
this case it is a gRPC stream that's closed because the context was
cancelled, so it will err out.

Closing the reading-end of the pipe side fixes this problem and unblocks
the command's `Wait()` call.

Fixes https://gitlab.com/gitlab-org/gitaly/-/issues/5953

[1]: https://gitlab.com/gitlab-org/gitaly/-/blob/faac94442a49a42427236f28d5190c6d721c172b/internal/gitaly/service/smarthttp/cache.go#L112
[2]: https://gitlab.com/gitlab-org/gitaly/-/blob/faac94442a49a42427236f28d5190c6d721c172b/internal/gitaly/service/smarthttp/cache.go#L92
[3]: https://gitlab.com/gitlab-org/gitaly/-/blob/faac94442a49a42427236f28d5190c6d721c172b/internal/gitaly/service/smarthttp/cache.go#L95-110
[4]: https://gitlab.com/gitlab-org/gitaly/-/blob/faac94442a49a42427236f28d5190c6d721c172b/internal/gitaly/service/smarthttp/cache.go#L104