Return ReferenceUpdateError from UserMergeBranch with transactions
UserMergeBranch is not returning an error detail when it fails due to
a concurrent reference update and transactions are enabled. The error
is different as the RPC handler has finished by the time we commit the
transaction and encounter the conflict. This commit adds handling for
this case in the transaction middleware so we conform to the API of
UserMergeBranch when committing the transaction fails due to a
concurrent reference update.