Bubble up the attempted update with reference verification errors
TransactionManager returns a ReferenceVerificationError when a
reference's old OID didn't match what was expected by the transaction
attempting to update the reference. In some contexts it would be
useful to have the new OID of the attempted update also available.
Include the new OID in the error so it can be bubbled up.