Use generated constants for RPC names in transaction middleware
We've copied the full names of various methods name in the
transaction middleware. Use the generated constants instead so we'll
have compilation failures if the RPC names change or the RPCs are
removed.

FetchObjectPool was previously typoed in the non-transactional list.
Now we actually don't start a transaction for it. Tests were adjusted
as the error messages are now the ones that are raised by the handler.