conflicts: Support CRLF line breaks in `ResolveConflicts` RPC
When parsing a conflicted file for the `ResolveConflicts` RPC, each line
is separated by a `\n` and patterned matched to look for conflict
markers. If the conflicted file uses CRLF line breaks, conflict markers
are not detected and thus no conflict resolution occurs.

Add support for conflicted files with CRLF line breaks.