backup: create directory if it does not exist
Prior to https://gitlab.com/gitlab-org/gitaly/-/merge_requests/6837,
the backup directory would be automatically created if it did not
exist.

We cannot use the `fileblob.CreateDir` option because this would make
the permissions of the directory world-readable and
world-writeable. Instead manually create the directory with the right
permissions.

Relates to https://gitlab.com/gitlab-org/charts/gitlab/-/issues/5467

Changelog: fixed