localrepo: Use `updateref.Update` for reference updates
In the `localrepo` package, we provide `Repo.UpdateRef`, which can be
used to update refs within a local repository. This command uses
'git-update-ref' to execute the update. Since we already have an
`updateref` package with better error handling, let's use that package
instead.

This causes the code to become a little verbose, since we need start,
commit, close the transaction. But this shouldn't affect perf, since Git
internally does that anyway when we run the 'git-update-ref' command.