referencetransaction: Notify transaction of default branch updates
Whenever the 'HEAD' symref is updated, the transaction manager should
know that it should track the default branch being changed. Once we move
to using the update-ref's symref commands, this information should be
conveyed to the transaction manager by the reference hook. So in the
hook, we now lookout for 'HEAD' symref updates and mark the default
branch as being updated.