Add consumer as a parameter to AcknowledgeTransaction
A `TransactionManager` may have multiple `LogConsumer`s registered, and
will need to track the position of each separately. Currently the
`AcknowledgeTransaction` method does not provide a way for the manager
to distinguish which consumer is notifying it.

Add the `LogConsumer` as a parameter to `AcknowledgeTransaction` so that
the manager can associate the call to that consumer.