Implement LogManager for TransactionManager
Implement the `LogManager` interface for `TransactionManager`, allowing
it to notify a `LogConsumer` of new log entries and prune entries when
the consumer has processed them. The manager stores the last
acknowledged LSN the consumer, but does not retain this state across
restarts. On restart, the low water mark will be set to zero as we
cannot assume any knowledge of the current state of the consumer.
The consumer's next acknowledgement will inform the manager of its
actual position.

To avoid injecting acknowledgements into the entire
`TransactionManager` test suite, we only add a `LogConsumer` to the
new consumer-related tests.