Fix TestTrackRepositoriesSubcommand flaky test
The flakiness happens if we hit an error here [1], which would cause
extra output to be present which we can't account for in the tests since
the chances of this error happening (could be a Gitaly connection
timeout) are very slim.

By testing against individual lines instead of the output as whole
string, we eliminate the source of the flakiness for this particular
test.

[1]:
https://gitlab.com/gitlab-org/gitaly/-/blob/7acf563cecc93e6c26c0c77151993e6ff0c17318/internal/cli/praefect/subcmd_track_repository.go#L352

Fixes https://gitlab.com/gitlab-org/gitaly/-/issues/5974