backup: Convert WAL archive worker count to uint
It is never valid for `LogEntryArchiver` to have a negative worker
count. We currently convert any value <= 0 to 1 in the constructor, but
for purposes of config validation it's better to make invalid values
impossible. By using a `uint` we can rely on `go-toml`'s parser to error
out on a negative value. No calculations are performed with the worker
count, so we have no risk of an overflow.

Convert `workerCount` and associated fields to `uint`.