Set Golang MAXPROCS
Despite having a way to set cpu limits on a container,
Go does not see the limits set on the container. GOMAXPROCS will
default to the cpu cores of the host-machine. Hence automaxprocs
is needed to automatically set the container's cpu limits. We
have also configured it to log using our Gitaly logger.